backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
opencv-python
numpy
ultralytics
